refresh in a nested router page error done and other changes
